// bora_no_standalone/src/app/interfaces/corredor/puestometro.ts
export interface Puestometro {
    carrera: string;
    dosmilveinte: number;
    dosmilveintiuno: number;
    dosmilveintidos: number;
    dosmilveintitres: number;
    dosmilveinticuatro: number;
    dosmilveinticinco: number;
    dosmilveintiseis: number;
    dosmilveintisiete: number;
    bandera: string;
    logo: string;
}